# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_projects__project_id__datasets_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_projects__project_id__datasets_upload_post",
        "type": "object"
      },
      "ColumnProfile": {
        "properties": {
          "dist": {
            "items": {
              "type": "number"
            },
            "title": "Dist",
            "type": "array"
          },
          "dtype": {
            "enum": [
              "int",
              "float",
              "category",
              "datetime",
              "bool",
              "string"
            ],
            "title": "Dtype",
            "type": "string"
          },
          "missing_pct": {
            "title": "Missing Pct",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "feature",
              "target",
              "id"
            ],
            "title": "Role",
            "type": "string"
          },
          "unique": {
            "title": "Unique",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "dtype",
          "role",
          "missing_pct",
          "unique",
          "dist"
        ],
        "title": "ColumnProfile",
        "type": "object"
      },
...
      "DatasetInfo": {
        "description": "What loading data returns: the stored version and its columns.",
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "data_version_id": {
            "title": "Data Version Id",
            "type": "string"
          },
          "default_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Target"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "short_hash": {
            "title": "Short Hash",
            "type": "string"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "data_version_id",
          "short_hash",
          "filename",
          "columns",
          "total_rows"
        ],
        "title": "DatasetInfo",
        "type": "object"
      },
...
      "FeatureSuggestion": {
        "properties": {
          "availability_assumption": {
            "default": "",
            "title": "Availability Assumption",
            "type": "string"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "llm": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider, model, cost and decision_mode of the call that proposed it",
            "title": "Llm"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "required_columns": {
            "items": {
              "type": "string"
            },
            "title": "Required Columns",
            "type": "array"
          },
          "risk": {
            "default": "",
            "title": "Risk",
            "type": "string"
          }
        },
        "required": [
          "name",
          "formula",
          "reason"
        ],
        "title": "FeatureSuggestion",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeakageFinding": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "check": {
            "title": "Check",
            "type": "string"
          },
          "column": {
            "title": "Column",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "column",
          "message",
          "severity",
          "category",
          "check",
          "evidence"
        ],
        "title": "LeakageFinding",
        "type": "object"
      },
...
      "SampleDatasetRequest": {
        "properties": {
          "dataset_name": {
            "default": "telecom_churn",
            "title": "Dataset Name",
            "type": "string"
          }
        },
        "title": "SampleDatasetRequest",
        "type": "object"
      },
      "SqlSnapshotRequest": {
        "description": "Run one read-only query and store the result as a data version.\n\nName a saved connection (``connection_id``, see /connections) or, until the UI moves to\nthose, pass a ``connection_string`` (never stored or echoed). Exactly one of the two.",
        "properties": {
          "connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A saved connection of this project",
            "title": "Connection Id"
          },
          "connection_string": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy: read-only connection; never stored or echoed",
            "title": "Connection String"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SqlSnapshotRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic ML experimentation platform",
    "title": "MLPilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/datasets/sample": {
      "post": {
        "description": "Load a bundled sample dataset as an immutable version.",
        "operationId": "load_sample_dataset_api_v1_projects__project_id__datasets_sample_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Sample Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/projects/{project_id}/datasets/sql": {
      "post": {
        "description": "Run one read-only query and store the result as an immutable version.",
        "operationId": "snapshot_sql_query_api_v1_projects__project_id__datasets_sql_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SqlSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Snapshot Sql Query",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/projects/{project_id}/datasets/upload": {
      "post": {
        "description": "Upload a CSV, store it as an immutable version and return its columns.",
        "operationId": "upload_dataset_api_v1_projects__project_id__datasets_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_projects__project_id__datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/projects/{project_id}/datasets/{data_version_id}/columns": {
      "get": {
        "operationId": "get_columns_api_v1_projects__project_id__datasets__data_version_id__columns_get",
        "parameters": [
          {
            "in": "path",
            "name": "data_version_id",
            "required": true,
            "schema": {
              "title": "Data Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_column",
            "required": true,
            "schema": {
              "title": "Target Column",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ColumnProfile"
                  },
                  "title": "Response Get Columns Api V1 Projects  Project Id  Datasets  Data Version Id  Columns Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Columns",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/projects/{project_id}/datasets/{data_version_id}/leakage": {
      "get": {
        "operationId": "get_leakage_api_v1_projects__project_id__datasets__data_version_id__leakage_get",
        "parameters": [
          {
            "in": "path",
            "name": "data_version_id",
            "required": true,
            "schema": {
              "title": "Data Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_column",
            "required": true,
            "schema": {
              "title": "Target Column",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LeakageFinding"
                  },
                  "title": "Response Get Leakage Api V1 Projects  Project Id  Datasets  Data Version Id  Leakage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Leakage",
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/datasets/{data_version_id}/suggestions": {
      "get": {
        "description": "Ask the planner for feature ideas (data only; nothing is run).",
        "operationId": "get_suggestions_api_v1_projects__project_id__datasets__data_version_id__suggestions_get",
        "parameters": [
          {
            "in": "path",
            "name": "data_version_id",
            "required": true,
            "schema": {
              "title": "Data Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_column",
            "required": true,
            "schema": {
              "title": "Target Column",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user_query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Query"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeatureSuggestion"
                  },
                  "title": "Response Get Suggestions Api V1 Projects  Project Id  Datasets  Data Version Id  Suggestions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Suggestions",
        "tags": [
          "datasets"
        ]
      }
    },
...
  }
}